+++ lib/kb_hmmer/kb_hmmerClient.py
@@ -83,7 +83,9 @@ class kb_hmmer(object):
            "input_many_ref" of type "data_obj_ref", parameter
            "output_filtered_name" of type "data_obj_name", parameter
            "coalesce_output" of type "bool", parameter "e_value" of Double,
-           parameter "bitscore" of Double, parameter "maxaccepts" of Double
+           parameter "bitscore" of Double, parameter "maxaccepts" of Double,
+           parameter "heatmap" of type "bool", parameter "vertical" of type
+           "bool", parameter "show_blanks" of type "bool"
         :returns: instance of type "HMMER_Output" (HMMER Output) ->
            structure: parameter "report_name" of type "data_obj_name",
            parameter "report_ref" of type "data_obj_ref"